Refactor wasm
