+++ guppylang-internals/src/guppylang_internals/engine.py
@@ -41,11 +41,11 @@ from guppylang_internals.error import (
     RequiresMonomorphizationError,
     pretty_errors,
 )
+from guppylang_internals.frame_util import get_calling_frame
 from guppylang_internals.metadata.debug_info_util import (
     StringTable,
 )
 from guppylang_internals.span import SourceMap
-from guppylang_internals.tracing.util import get_calling_frame
 from guppylang_internals.tys.arg import ConstArg, TypeArg
 from guppylang_internals.tys.builtin import (
     array_type_def,
@@ -495,7 +495,6 @@ class CompilationEngine:
         # Set up string tables for metadata serialization. We know that the first entry
         # in the table is always the file containing the Hugr entrypoint.
         frame = get_calling_frame()
-        assert frame is not None
         filename = frame.f_code.co_filename
 
         ctx = CompilerContext(graph, set(def_ids), StringTable())
