+++ guppylang-internals/src/guppylang_internals/tracing/util.py
@@ -1,11 +1,11 @@
 import functools
-import inspect
 import sys
 from collections.abc import Callable
-from types import FrameType, ModuleType, TracebackType
+from types import TracebackType
 from typing import ParamSpec, TypeVar
 
 from guppylang_internals.error import GuppyComptimeError, GuppyError, exception_hook
+from guppylang_internals.frame_util import remove_internal_frames
 
 P = ParamSpec("P")
 T = TypeVar("T")
@@ -76,32 +76,3 @@ def tracing_except_hook(
         ipython_shell.excepthook(excty, err, traceback)
     except NameError:
         sys.__excepthook__(excty, err, traceback)
-
-
-def get_calling_frame() -> FrameType | None:
-    """Finds the first frame that called this function outside the compiler."""
-    frame = inspect.currentframe()
-    while frame:
-        module = inspect.getmodule(frame)
-        if module is None or not is_compiler_module(module):
-            return frame
-        frame = frame.f_back
-    return None
-
-
-def remove_internal_frames(tb: TracebackType | None) -> TracebackType | None:
-    """Removes internal frames relating to the Guppy compiler from a traceback."""
-    if tb:
-        module = inspect.getmodule(tb.tb_frame)
-        if module is not None and is_compiler_module(module):
-            return remove_internal_frames(tb.tb_next)
-        if tb.tb_next:
-            tb.tb_next = remove_internal_frames(tb.tb_next)
-    return tb
-
-
-def is_compiler_module(module: ModuleType) -> bool:
-    """Checks whether a given Python module belongs to the Guppy compiler."""
-    return module.__name__.startswith(
-        "guppylang_internals."
-    ) or module.__name__.startswith("guppylang.")
